Add script to add information to the tile grid
Some information like the size of the plb and pib tiles has been deduced
from looking at repeating patterns in the locations of such tiles.

Later it was discovered that the clock spines are actually 8 bits high
in the bitstream rather than the 2 bits implied by the bl_beg numbers.

This requires ofsetting the start of each tile after these by 6 bits to
ensure they are located correctly.  These locations may need tweaking
per eagle device but this can be done in the future