Add adapted fuzzing helpers from libtrellis
The fuzzing infrastructure is mostly imported from Project Trellis but
has required some adaption for toolchain differences and the routing
fuzzer is much less automatic than the Trellis one but allows checking
if a given arc in a design sets any bits and adding those to the db.

A catch-all template tcl file is also added for the general case of
importing a pnl file and producing a bitstream.

The fuzzing helpers directory is also added to PYTHONPATH in
environment.sh